Resolved migration issues
