+++ pages/models.py
@@ -1,26 +1,3 @@
 from django.db import models
-from django.conf import settings
-from django.shortcuts import reverse
-from django_countries.fields import CountryField
-
 
 # Create your models here.
-class CheckoutAddress(models.Model):
-    user = models.ForeignKey(settings.AUTH_USER_MODEL, on_delete=models.CASCADE)
-    street_address = models.CharField(max_length=100)
-    apartment_address = models.CharField(max_length=100)
-    country = CountryField(multiple=False)
-    zip = models.CharField(max_length=100)
-
-    def __str__(self):
-        return self.user.username
-
-class Payment(models.Model):
-    stripe_id = models.CharField(max_length=50)
-    user = models.ForeignKey(settings.AUTH_USER_MODEL, 
-                            on_delete=models.SET_NULL, blank=True, null=True)
-    amount = models.FloatField()
-    timestamp = models.DateTimeField(auto_now_add=True)
-
-    def __str__(self):
-        return self.user.username
