+++ payment/forms.py
@@ -2,11 +2,6 @@ from django import forms
 from django_countries.fields import CountryField
 from django_countries.widgets import CountrySelectWidget
 
-PAYMENT = (
-    ('S', 'Stripe'),
-    ('P', 'Paypal')
-)
-
 class CheckoutForm(forms.Form):
     street_address = forms.CharField(widget=forms.TextInput(attrs={
         'class':'form-control',
@@ -20,7 +15,4 @@ class CheckoutForm(forms.Form):
     }))
     same_billing_address = forms.BooleanField(required=False)
     save_info = forms.BooleanField(required=False)
-    payment_option = forms.ChoiceField(
-        widget=forms.RadioSelect, choices=PAYMENT
-    )
