+++ requirements.txt
@@ -1,7 +1,9 @@
 asgiref==3.5.2
 Django==4.1.3
+django-countries==7.5
 django-crispy-forms==1.14.0
 django-environ==0.9.0
 Pillow==9.3.0
 sqlparse==0.4.3
+typing_extensions==4.4.0
 whitenoise==6.2.0
